requirements.txt

flask>=3.0.0
pandas>=2.2.0
openpyxl>=3.1.2
numpy>=1.26.0
plotly>=5.18.0
sqlalchemy>=2.0.0
streamlit>=1.31.0
scikit-learn>=1.4.0
scipy>=1.12.0
seaborn>=0.13.0
python-pptx>=0.6.22
